requirements.txt
streamlit==1.32.2
requests==2.31.0
pyOpenSSL==24.0.0
cryptography==42.0.5
python-dotenv==1.0.1 